+++ packages/ui/src/api/newsletter/create-subscription.ts
@@ -15,28 +15,9 @@ export type TnewsletterSubscriptionState = {
   status: TnewsletterSubscriptionStatus
 }
 
-// Function to verify reCAPTCHA token using Google's siteverify API
-async function fnVerifyRecaptcha(iToken: string): Promise<boolean> {
-  const LSecretKey = process.env.RECAPTCHA_SECRET_KEY
-  const LRecaptchaUrl = `https://www.google.com/recaptcha/api/siteverify?secret=${LSecretKey}&response=${iToken}`
-
-  try {
-    const LdResponse = await fetch(LRecaptchaUrl, { method: "POST" })
-    const LdData = await LdResponse.json()
-
-    // Return true only if verification is successful and the score is above threshold
-    return LdData.success && LdData.score >= 0.5
-    
-  } catch (error) {
-    console.error("reCAPTCHA verification error:", error)
-    return false
-  }
-}
-
 // Define expected structure and constraints for form data
 const LdSchema = z.object({
   email: z.string().email("Please enter a valid email"),
-  recaptchaToken: z.string(),
 })
 
 export async function subscribeNewsletter(
@@ -53,17 +34,13 @@ export async function subscribeNewsletter(
   // Validate form input early to fail fast and give user feedback before hitting the backend
   const LdParsed = LdSchema.safeParse({
     email: idFormData.get("email"),
-    recaptchaToken: idFormData.get("recaptchaToken"),
   })
 
   // If validation fails, return a user-friendly message without making any API requests
   if (!LdParsed.success) {
-    const LdErrors = LdParsed.error.flatten().fieldErrors
     return {
       message:
-        LdErrors.email?.[0] ??
-        LdErrors.recaptchaToken?.[0] ??
-        "Invalid input",
+        LdParsed.error.flatten().fieldErrors.email?.[0] || "Invalid input",
       status: "error",
     }
   }
@@ -73,13 +50,6 @@ export async function subscribeNewsletter(
 
   // Disables TLS verification for local development
   // process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0"
-
-  // Verify if the form submission is made by a human
-  const LIsHuman = await fnVerifyRecaptcha(LdParsed.data.recaptchaToken,)
-  if (!LIsHuman) {
-    return {  message: "reCAPTCHA verification failed",
-      status: "error",}
-  }
 
   try {
     // First check: is this email already subscribed? Prevents duplicate entries and unnecessary API calls
