+++ packages/ui/src/components/subscription.tsx
@@ -2,14 +2,12 @@
 
 import { Button } from "@repo/ui/components/ui/button";
 import { Input } from "@repo/ui/components/ui/input";
-import {
-  subscribeNewsletter,
-  type TnewsletterSubscriptionState,
-} from "@repo/ui/api/newsletter/create-subscription";
 import { TNewsletterSubscriptionProps } from "@repo/middleware/types";
 import posthog from "posthog-js";
 import { useEffect, useState } from "react";
 import { ReCaptchaProvider, useReCaptcha } from "next-recaptcha-v3"
+import { fnSubscribewithCaptcha } from "@repo/ui/api/newsletter/subscription-captcha";
+import { TnewsletterSubscriptionState } from "@repo/ui/api/newsletter/create-subscription";
 
 const LdVariants = {
   sm: {
@@ -111,7 +109,7 @@ export function NewsletterSubscriptionForm({
         });
       }
       // subscription apis
-      const LdResult = await subscribeNewsletter(
+      const LdResult = await fnSubscribewithCaptcha(
         { message: "", status: "error" },
         LdFormData,
       );
